backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
google-generativeai
pdfplumber
python-multipart
pgvector
requests
